feat(STONEINTG-674): update kubebuilder to v4
update kubebuilder to v4 according to https://book.kubebuilder.io
/migration/manually_migration_guide_gov3_to_gov4
* move main.go to cmd folder
* rename contollers folder to controller under  new internal folder
* update Makefile
* update Dockerfile
* update admissionv1beta1 to admissionv1
* update PROJEDT
