+++ Makefile
@@ -123,11 +123,11 @@ test: manifests generate fmt vet envtest download-crds ## Run tests.
 
 .PHONY: build
 build: generate fmt vet ## Build manager binary.
-	go build -o bin/manager main.go
+	go build -o bin/manager cmd/main.go
 
 .PHONY: run
 run: manifests generate fmt vet ## Run a controller from your host.
-	go run ./main.go
+	go run ./cmd/main.go
 
 .PHONY: img-build
 img-build: test ## Build container image with the manager.
