+++ internal/controller/component/component_adapter.go
