+++ internal/controller/component/component_adapter_test.go
