+++ internal/controller/integrationpipeline/integrationpipeline_adapter.go
