+++ internal/controller/scenario/scenario_adapter.go
