+++ internal/controller/scenario/scenario_adapter_test.go
