+++ internal/controller/snapshot/snapshot_adapter.go
