+++ internal/controller/statusreport/statusreport_adapter.go
