+++ internal/controller/statusreport/statusreport_adapter_test.go
